+++ opengdflow-back/src/main/java/br/com/opengd/controller/response/DocumentoResponse.java
@@ -3,6 +3,7 @@ package br.com.opengd.controller.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -15,7 +16,7 @@ public class DocumentoResponse {
     private UsuarioDadosPrimariosResponse usuario;
     private String nome;
     private String descricao;
-    private Long tamanho;
+    private BigDecimal tamanho;
     private String tipo;
     private LocalDateTime dataHora;
 }
